cloud/project/backend/requirements.txt
fastapi
uvicorn
google-cloud-firestore
google-cloud-translate
google-cloud-storage
google-cloud-texttospeech
google-cloud-pubsub
python-multipart
python-dotenv
